feat: support tags filtering in other filesystem mcp tools #4863
