+++ front/lib/actions/mcp_internal_actions/servers/data_warehouses/helpers.ts
@@ -27,9 +27,9 @@ export async function getAvailableWarehouses(
   >
 > {
   const coreAPI = new CoreAPI(config.getCoreAPIConfig(), logger);
-  const dataSourceViewFilter = makeCoreSearchNodesFilters(
-    dataSourceConfigurations
-  ).map((view) => ({
+  const dataSourceViewFilter = makeCoreSearchNodesFilters({
+    agentDataSourceConfigurations: dataSourceConfigurations,
+  }).map((view) => ({
     ...view,
     search_scope: "data_source_name" as const,
   }));
@@ -153,7 +153,9 @@ export async function getWarehouseNodes(
   const result = await coreAPI.searchNodes({
     query,
     filter: {
-      data_source_views: makeCoreSearchNodesFilters(configsToUse),
+      data_source_views: makeCoreSearchNodesFilters({
+        agentDataSourceConfigurations: configsToUse,
+      }),
       parent_id: parentIdToUse ?? undefined,
       node_ids: nodeIdsToUse,
     },
@@ -316,7 +318,9 @@ export async function validateTables(
   const coreAPI = new CoreAPI(config.getCoreAPIConfig(), logger);
   const searchResult = await coreAPI.searchNodes({
     filter: {
-      data_source_views: makeCoreSearchNodesFilters([relevantConfig]),
+      data_source_views: makeCoreSearchNodesFilters({
+        agentDataSourceConfigurations: [relevantConfig],
+      }),
       node_ids: parsedTables.map((t) => t.nodeId),
     },
   });
